+++ otherlibs/stdune/path.mli
@@ -98,6 +98,8 @@ module Source : sig
 
   val of_local : Local.t -> t
 
+  (** [relative dir s] if s can be ".." it could escape the working directory.
+      {!Path.relative} should be used instead. *)
   val relative : ?error_loc:Loc0.t -> t -> string -> t
 
   val split_first_component : t -> (string * Local.t) option